implement the new client interface
Add `bus.getProxyObject()` to get a proxy object for an object on the
bus. The proxy object contains introspection data for the object
including the nodes and interfaces.

Interfaces can be gotten with `object.getInterface()` with the name of
the interface. The interface includes methods to call on the bus based
on the introspection data.

The interface is an event emitter which emits signals for the object on
the bus.

See the new mpris.js example and the README for more details.